Update the 'Source Code' block titles to 'Pseudo Code'.
The name 'Source Code' is a bit misleading as they are not actual dumps of real code. They are pseudo code generated from Ghidra and IDA, then hand-rewritten to be human readable. They also do not contain every exact detail from all handlers. They are stripped down and rewritten to be easiest to understand.